Move controllers that are not actual SECOM Controllers
